Adding computation of total br length (comput_anv -> note: name to be changed)
In phymat: adding comput_anv function. Compute the total branch lenght to find anchor for absolute node values and ultrametry
In shfunc: minor corrections to admit multiple trees
In TTiP: adding comput_anv. Modification in objects, but to be removed
In verif: adding an object to be returned. But probably to be removed